hw9 - task 1

+++ src/test/java/edu/hw9/task1/StatsCollectorTest.java
@@ -2,7 +2,6 @@ package edu.hw9.task1;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import java.util.HashMap;
 import java.util.Map;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
@@ -17,14 +16,14 @@ class StatsCollectorTest {
 
     @Test
     void collectorTest() throws InterruptedException {
-        collector.push("metric_name", new double[] {0.1, 0.05, 1.4, 5.1, 0.3});
+        collector.push("metric_name", new float[] {0.1f, 0.05f, 1.4f, 5.1f, 0.3f});
         Thread.sleep(100);
 
-        Map<String, Double> expected = Map.of(
-            "metric_name_sum", 6.95,
-            "metric_name_average", 1.39,
-            "metric_name_min", 0.05,
-            "metric_name_max", 5.1
+        Map<String, Float> expected = Map.of(
+            "metric_name_sum", 6.95f,
+            "metric_name_average", 1.39f,
+            "metric_name_min", 0.05f,
+            "metric_name_max", 5.1f
         );
 
         assertThat(collector.stats()).isEqualTo(expected);
